multidoc_search/gemini/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu==1.7.4
langchain_google_genai
python-docx
openpyxl
altair<5
